+++ src/olmo_core/model_ladder.py
@@ -322,21 +322,21 @@ class ModelLadder(Config, metaclass=ABCMeta):
         :param gpu_type: The type of GPU as given by ``torch.cuda.get_device_name()``.
         :param dp_world_size: The data parallel world size.
         """
-        del gpu_type, dp_world_size
+        del dp_world_size
 
         from olmo_eval import list_tasks
 
-        return (
-            TrainerConfig(
+        config = TrainerConfig(
                 save_folder=self.get_save_folder(size),
                 metrics_collect_interval=10,
-                cancel_check_interval=1,
+                cancel_check_interval=10,
                 max_duration=self.get_duration(size),
             )
-            .with_callback("gpu_monitor", GPUMemoryMonitorCallback())
-            .with_callback("config_saver", ConfigSaverCallback())
-            .with_callback("garbage_collector", GarbageCollectorCallback())
-            .with_callback(
+        if gpu_type not in ("cpu", "mps"):
+            config = config.with_callback("gpu_monitor", GPUMemoryMonitorCallback())
+        config = config.with_callback("config_saver", ConfigSaverCallback())
+        config = config.with_callback("garbage_collector", GarbageCollectorCallback())
+        config = config.with_callback(
                 "lm_evaluator",
                 LMEvaluatorCallbackConfig(
                     eval_dataset=NumpyDatasetConfig.from_data_mix(
@@ -350,7 +350,7 @@ class ModelLadder(Config, metaclass=ABCMeta):
                     eval_interval=500,
                 ),
             )
-            .with_callback(
+        config = config.with_callback(
                 "downstream_evaluator",
                 DownstreamEvaluatorCallbackConfig(
                     tasks=[
@@ -360,7 +360,7 @@ class ModelLadder(Config, metaclass=ABCMeta):
                     eval_interval=500,
                 ),
             )
-            .with_callback(
+        config = config.with_callback(
                 "checkpointer",
                 CheckpointerCallback(
                     save_interval=100_000,  # large enough value that we won't save until the end
@@ -368,7 +368,7 @@ class ModelLadder(Config, metaclass=ABCMeta):
                     save_async=True,
                 ),
             )
-            .with_callback(
+        config = config.with_callback(
                 "comet",
                 CometCallback(
                     name=f"{self.name}-{size}",
@@ -378,7 +378,7 @@ class ModelLadder(Config, metaclass=ABCMeta):
                     cancel_check_interval=5,
                 ),
             )
-            .with_callback(
+        config = config.with_callback(
                 "wandb",
                 WandBCallback(
                     name=f"{self.name}-{size}",
@@ -388,7 +388,7 @@ class ModelLadder(Config, metaclass=ABCMeta):
                     cancel_check_interval=5,
                 ),
             )
-        )
+        return config
 
     def validate(self):
         """